Add files via upload
PASSWORD GENERATOR

This is a simple application which creates a random password based on the user given length of password using python
Here, The input is given by user with fixed length and complexity for password and with the use of Random module we are going design password with alphabets, digits and special symbols. Therefore it generates randomly  password.